Fix BaseHead tests
